Completed GetItemsSold function for the dashboard

+++ BeanBag/Services/IDashboardAnalyticsService.cs
@@ -10,7 +10,7 @@ namespace BeanBag.Services
         public int GetTotalItems(string id);
         public IQueryable GetTopItems(string id);
         int GetItemsAvailable(string id);
-        public int GetItemsSold(string id);
+        public double GetItemsSold(string id);
 
     }
 }
